fix QTY field in order to make it float

+++ src/backend/models/models.py
@@ -107,7 +107,7 @@ class OrderItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
-    quantity = Column(Integer, nullable=False)
+    quantity = Column(Float, nullable=False)
     price = Column(Numeric, nullable=False)
     dispatched = Column(Boolean, default=False, nullable=False)
